min agent demands without breaking build
