+++ ci/build-single-steps.yml
@@ -20,7 +20,11 @@ steps:
   displayName: npm install
 
 # Verify min agent version demands
-- script: npm run verifyMinAgentDemands
+- script: |
+    cd ci
+    cd verifyMinAgentDemands
+    npm install
+    node index.js
   displayName: Verify all min agent demands are valid
 
 # Build
