+++ ci/verifyMinAgentDemands/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "verifyminagentdemands",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@octokit/rest": "^17.0.0",
+    "azure-pipelines-task-lib": "^2.9.3",
+    "semver": "4.3.3"
+  }
+}
